requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
google-cloud-vision==3.5.0 